se mejoro el metodo getMayorVenta para que sume bien la cantidad de productos.

+++ src/main/java/com/tpintegrador/bazar/model/dto/MayorVentaDto.java
@@ -7,7 +7,7 @@ public class MayorVentaDto {
 
     private Long codigoVenta;
     private Double total;
-    private Double cantidadProductos;
+    private Long cantidadProductos;
     private String nombreCliente;
     private String apellidoCliente;
 }
